// src/app/modules/core/interfaces/embarcacion.interface.ts
export interface Embarcacion{
    id?: string;
    capitan_barco?: string;
    embarcacion_nombre?: string;
    numeroRegistro?: string;
    puertoMarina?: string;
    embarcacionTipo?: string;
    marca?: string;
    tonelaje?: string;
    modelo?: string;
    year?: string;
    size?: string;
    numeroPasajeros?: string;
    urlBarco?: string;
    clasificacion?: string;
    nivelSeguridad?: string;
    country?: string;
    countryFlag?: string;
    lat?: string;
    lng?: string;
    tripulacion?: string;
    Viajes?: any[];
 

}