// src/app/modules/core/interfaces/detalle.interface.ts
export interface Detalle{
    id?: string;
    capitan_barco?: string;
    embarcacion_nombre?: string;
    numeroPasajeros?: any;
    destino?: string;
    numeroDias?: string;
    experiencia?: string;
    precio?: string;
    embarcacion_id?: string;
    urlFoto?: string;
    origen?: string;
    preciodesc?: any;
    tripulacion?: any;
    pais?: any;
    nombreCapitan?: any;
    urlCapitan?: any;
    tipoDetalle?: any;
    viaje_id?:any;
    viajesEmbarcacion?:any;
    embarcacionNombre?:any;
    lat?:any;
    lng?:any;
}